ensure scroll percentage is calculated correctly
Prior to this commit, under certain conditions it was impossible for the
scrollPercentageToDisplay to ever trigger.

Consider the following:
- page content height of 800px
- scrollPercentageToDisplay setting of 30 (percent)
- window height of 600px

In the above scenario, the user would need to scroll 240 pixels down the
page to trigger a modal, however there is only 200px of available
scrolling. The modal will never trigger.

This commit changes the arithmetic of percentageScrolled such that in
the above scenario the modal will correctly trigger after scrolling 30%
of the available space.